keepalived: set default run directory for pid file on build
This fixes a runtime startup error on system which does not have a
toplevel runtime directory for the pid file. On openwrt the pid is
located at /var/run and not on /run. To fix that add a configure option to
move the pid location to /var/run.

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit bc98aaa3f711f1cd6f7fa3a65da43411c2db98aa)

* Patch adapted for the branch so that it can be applied correctly